Refactorizacion del crear producto.

+++ Models/Productos.cs
@@ -4,5 +4,5 @@ public class Productos
 {
     public int idProducto { get; set; }
     public string Descripcion { get; set; }
-    public int Precio { get; set; }
+    public double Precio { get; set; }
 }
